VIEW PORT BUG WHEN SLIDING.
something like this:
https://stackoverflow.com/questions/57686840/why-does-transform-translatex-affect-a-fixed-elements-height-on-mobile

when the card is sliding, if you touch
the screen on mobile, your are "picking" a spot on teh screen.
when the card slides the viewport changes size.

I think somehow with a touch and a translate,
the mobile version treats that as a stretch?

I don't know, but the final fix was
a very rude "don't let th user zoom".