rework the flipping code.
it is too complicated.

too much state.

the goal is to make it so that if you click the button, or swip before the
back side is displayed, clicking wil flip the card.
sliding will slowly rotate the card.  and it will flip over/move back
depending on how far you have slid your finger.